event put, post

+++ src/main/java/com/calendar/models/User.java
@@ -20,7 +20,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    @Getter private Integer id;
+    @Getter private Long id;
 
     @Column
     private String login;
