+++ src/main/java/com/calendar/repositories/UserRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
-    User findUserById(Integer id);
+public interface UserRepository extends JpaRepository<User, Long> {
+    User findUserById(Long id);
 
     User findUserByLogin(String login);
